BE: create-pricebook-api

+++ backend/Feartures/Pricebooks/Create/CreatePricebookRequest.cs
@@ -6,8 +6,8 @@ namespace backend.Feartures.Pricebooks.Create
     {
         public long ProductId { get; set; }
         public string Name { get; set; }
-        public long MsrpPrice { get; set; }
-        public long FloorPrice { get; set; }
+        public decimal MsrpPrice { get; set; }
+        public decimal FloorPrice { get; set; }
         public DateTime Effective_To { get; set; }
         public PricebookStatus Status { get; set; } = PricebookStatus.Actived;
